# Expose size as an upper bound in list operations