--- a/src/pps/pps.proto
+++ b/src/pps/pps.proto
@@ -187,8 +187,8 @@ message ProcessStats {
   google.protobuf.Duration download_time = 1;
   google.protobuf.Duration process_time = 2;
   google.protobuf.Duration upload_time = 3;
-  uint64 download_bytes = 4;
-  uint64 upload_bytes = 5;
+  int64 download_bytes = 4;
+  int64 upload_bytes = 5;
 }
 
 message AggregateProcessStats {