--- a/src/server/pfs/server/driver.go
+++ b/src/server/pfs/server/driver.go
@@ -242,21 +242,21 @@ func (d *driver) inspectRepo(txnCtx *txncontext.TransactionContext, repo *pfs.Re
 	if repo == nil {
 		return nil, errors.New("repo cannot be nil")
 	}
-	result := &pfs.RepoInfo{}
-	if err := d.repos.ReadWrite(txnCtx.SqlTx).Get(pfsdb.RepoKey(repo), result); err != nil {
+	repoInfo := &pfs.RepoInfo{}
+	if err := d.repos.ReadWrite(txnCtx.SqlTx).Get(pfsdb.RepoKey(repo), repoInfo); err != nil {
 		return nil, err
 	}
 	if includeAuth {
 		permissions, roles, err := d.getPermissions(txnCtx.ClientContext, repo)
 		if err != nil {
 			if auth.IsErrNotActivated(err) {
-				return result, nil
+				return repoInfo, nil
 			}
 			return nil, errors.Wrapf(grpcutil.ScrubGRPC(err), "error getting access level for %q", repo)
 		}
-		result.AuthInfo = &pfs.RepoAuthInfo{Permissions: permissions, Roles: roles}
+		repoInfo.AuthInfo = &pfs.RepoAuthInfo{Permissions: permissions, Roles: roles}
 	}
-	return result, nil
+	return repoInfo, nil
 }
 
 func (d *driver) getPermissions(ctx context.Context, repo *pfs.Repo) ([]auth.Permission, []string, error) {
@@ -275,11 +275,11 @@ func (d *driver) listRepo(ctx context.Context, includeAuth bool, repoType string
 	repoInfo := &pfs.RepoInfo{}
 
 	processFunc := func(string) error {
-		size, err := d.getRepoSize(ctx, repoInfo.Repo)
+		size, err := d.repoSizeUpperBound(ctx, repoInfo.Repo)
 		if err != nil {
 			return err
 		}
-		repoInfo.SizeBytes = uint64(size)
+		repoInfo.SizeBytesUpperBound = size
 		if includeAuth && authSeemsActive {
 			permissions, roles, err := d.getPermissions(ctx, repoInfo.Repo)
 			if err == nil {
@@ -679,7 +679,7 @@ func (d *driver) aliasCommit(txnCtx *txncontext.TransactionContext, parent *pfs.
 	return commitInfo, nil
 }
 
-func (d *driver) getRepoSize(ctx context.Context, repo *pfs.Repo) (int64, error) {
+func (d *driver) repoSizeUpperBound(ctx context.Context, repo *pfs.Repo) (int64, error) {
 	repoInfo := new(pfs.RepoInfo)
 	if err := d.repos.ReadOnly(ctx).Get(pfsdb.RepoKey(repo), repoInfo); err != nil {
 		return 0, err
@@ -690,7 +690,24 @@ func (d *driver) getRepoSize(ctx context.Context, repo *pfs.Repo) (int64, error)
 			if err := d.branches.ReadOnly(ctx).Get(pfsdb.BranchKey(branch), branchInfo); err != nil {
 				return 0, err
 			}
-			return d.sizeOfCommit(ctx, branchInfo.Head)
+			return d.commitSizeUpperBound(ctx, branchInfo.Head)
+		}
+	}
+	return 0, nil
+}
+
+func (d *driver) repoSize(ctx context.Context, repo *pfs.Repo) (int64, error) {
+	repoInfo := new(pfs.RepoInfo)
+	if err := d.repos.ReadOnly(ctx).Get(pfsdb.RepoKey(repo), repoInfo); err != nil {
+		return 0, err
+	}
+	for _, branch := range repoInfo.Branches {
+		if branch.Name == "master" {
+			branchInfo := &pfs.BranchInfo{}
+			if err := d.branches.ReadOnly(ctx).Get(pfsdb.BranchKey(branch), branchInfo); err != nil {
+				return 0, err
+			}
+			return d.commitSize(ctx, branchInfo.Head)
 		}
 	}
 	return 0, nil
@@ -913,14 +930,14 @@ func (d *driver) inspectCommit(ctx context.Context, commit *pfs.Commit, wait pfs
 	}
 
 	if commitInfo.Finished != nil {
-		size, err := d.sizeOfCommit(ctx, commitInfo.Commit)
+		size, err := d.commitSize(ctx, commitInfo.Commit)
 		if err != nil {
 			return nil, err
 		}
 		if commitInfo.Details == nil {
 			commitInfo.Details = &pfs.CommitInfo_Details{}
 		}
-		commitInfo.Details.SizeBytes = uint64(size)
+		commitInfo.Details.SizeBytes = size
 	}
 	return commitInfo, nil
 }
@@ -1053,7 +1070,7 @@ func (d *driver) getCommit(ctx context.Context, commit *pfs.Commit) (*pfs.Commit
 	return commitInfo, nil
 }
 
-func (d *driver) listCommit(ctx context.Context, repo *pfs.Repo, to *pfs.Commit, from *pfs.Commit, number uint64, reverse bool, cb func(*pfs.CommitInfo) error) error {
+func (d *driver) listCommit(ctx context.Context, repo *pfs.Repo, to *pfs.Commit, from *pfs.Commit, number int64, reverse bool, cb func(*pfs.CommitInfo) error) error {
 	// Validate arguments
 	if repo == nil {
 		return errors.New("repo cannot be nil")
@@ -1090,7 +1107,7 @@ func (d *driver) listCommit(ctx context.Context, repo *pfs.Repo, to *pfs.Commit,
 
 	// if number is 0, we return all commits that match the criteria
 	if number == 0 {
-		number = math.MaxUint64
+		number = math.MaxInt64
 	}
 
 	if from != nil && to == nil {
@@ -1110,6 +1127,11 @@ func (d *driver) listCommit(ctx context.Context, repo *pfs.Repo, to *pfs.Commit,
 
 				if reverse {
 					ci = cis[len(cis)-1-i]
+				}
+				var err error
+				ci.SizeBytesUpperBound, err = d.commitSizeUpperBound(ctx, ci.Commit)
+				if err != nil {
+					return err
 				}
 				if err := cb(ci); err != nil {
 					return err