--- a/src/server/pfs/server/option.go
+++ b/src/server/pfs/server/option.go
@@ -6,18 +6,7 @@ import "github.com/pachyderm/pachyderm/v2/src/internal/storage/fileset"
 type SourceOption func(*sourceConfig)
 
 type sourceConfig struct {
-	details bool
-	filter  func(fileset.FileSet) fileset.FileSet
-}
-
-// WithDetails sets the source to populate the 'Details' field of the returned
-// files.  This means resolving the indexes for computing the true hashes of the
-// files, and computing the directory hashes / sizes by scanning ahead in a
-// parallel iteration.
-func WithDetails() SourceOption {
-	return func(sc *sourceConfig) {
-		sc.details = true
-	}
+	filter func(fileset.FileSet) fileset.FileSet
 }
 
 // WithFilter applies a filter to the fileset after it has been set up by the source.